requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==24.2.0
blinker==1.9.0
blobfile==3.0.0
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
-e git+https://github.com/CompVis/zigma.git@f235fc801ec863aec895b42457d5e9e5e5c09d5c#egg=causal_conv1d&subdirectory=dis_causal_conv1d
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
clip @ file:///home/cr7/CLIP-main
cloudpickle==3.1.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.40.0
decorator==4.4.2
Deprecated==1.2.15
diffusers==0.31.0
docker==7.1.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
et_xmlfile==2.0.0
executing==2.1.0
fairscale==0.4.13
filelock @ file:///croot/filelock_1700591183607/work
Flask==3.1.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
-e git+https://github.com/Baixuzx7/TeRF.git@1db2d463fdd1a521061e5306f594c203d24ae75d#egg=groundingdino
grpcio==1.68.1
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.26.2
hydra-core==1.3.2
icecream==2.1.3
idna @ file:///croot/idna_1714398848350/work
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
inplace-abn==1.1.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
lightning-utilities==0.11.8
loguru==0.7.2
lxml==5.3.0
Mako==1.3.8
-e git+https://github.com/CompVis/zigma.git@f235fc801ec863aec895b42457d5e9e5e5c09d5c#egg=mamba_ssm&subdirectory=dis_mamba
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
memory-profiler==0.61.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml_collections==0.1.1
mlflow==2.19.0
mlflow-skinny==2.19.0
mmcv-full==1.7.2
moviepy==1.0.3
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
networkx @ file:///croot/networkx_1717597493534/work
ninja==1.11.1.1
nose==1.3.7
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
nvidia-ml-py==12.535.161
nvitop==1.3.2
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.1
pandas==2.2.3
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.3.6
portalocker==3.0.0
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycryptodomex==3.21.0
Pygments==2.18.0
pyparsing==3.2.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-wavelets==1.3.0
pytorch_toolbelt==0.8.0
pytz==2024.2
PyWavelets==1.7.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2024.9.11
requests @ file:///croot/requests_1716902831423/work
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tokenizers==0.20.1
torch==2.3.1
torch-encoding @ git+https://github.com/zhanghang1989/PyTorch-Encoding/@96e71b025876ee2df1dc0fc805cab12a69af05b1
torch-fidelity==0.3.0
torchdiffeq==0.2.4
torchmetrics==1.5.1
torchvision==0.18.1
tqdm==4.66.6
transformers==4.46.1
triton==2.3.1
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1715635851070/work
Wand==0.6.13
wandb==0.18.5
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
wrapt==1.17.0
XlsxWriter==3.2.0
yacs==0.1.8
yapf==0.43.0
yarl==1.16.0
zipp==3.20.2
